refactor: move subject file functionality from annif.corpus to annif.vocab submodule
