+++ annif/corpus/__init__.py
@@ -8,19 +8,13 @@ from .document import (
     LimitingDocumentCorpus,
     TransformingDocumentCorpus,
 )
-from .skos import SubjectFileSKOS
-from .subject import Subject, SubjectFileCSV, SubjectFileTSV, SubjectSet
-from .types import Document
+from .types import Document, SubjectSet
 
 __all__ = [
     "DocumentDirectory",
     "DocumentFile",
     "DocumentList",
-    "Subject",
-    "SubjectFileTSV",
-    "SubjectFileCSV",
     "SubjectSet",
-    "SubjectFileSKOS",
     "Document",
     "CombinedCorpus",
     "TransformingDocumentCorpus",
