+++ annif/vocab/__init__.py
@@ -1,11 +1,18 @@
 """Annif vocabulary functionality"""
 
+from .skos import SubjectFileSKOS
+from .subject_file import SubjectFileCSV, SubjectFileTSV
 from .subject_index import SubjectIndexFile, SubjectIndexFilter
-from .types import SubjectIndex
+from .types import Subject, SubjectCorpus, SubjectIndex
 from .vocab import AnnifVocabulary
 
 __all__ = [
     "AnnifVocabulary",
+    "Subject",
+    "SubjectCorpus",
+    "SubjectFileCSV",
+    "SubjectFileSKOS",
+    "SubjectFileTSV",
     "SubjectIndex",
     "SubjectIndexFile",
     "SubjectIndexFilter",
