+++ annif/vocab/types.py
@@ -3,8 +3,32 @@
 from __future__ import annotations
 
 import abc
+import collections
 
-from annif.corpus import Subject
+Subject = collections.namedtuple("Subject", "uri labels notation")
+
+
+class SubjectCorpus(metaclass=abc.ABCMeta):
+    """Abstract base class for subject corpora"""
+
+    @property
+    @abc.abstractmethod
+    def subjects(self):
+        """Iterate through the subject corpus, yielding Subject objects."""
+        pass  # pragma: no cover
+
+    @property
+    @abc.abstractmethod
+    def languages(self):
+        """Provide a list of language codes supported by this subject
+        corpus."""
+        pass  # pragma: no cover
+
+    @abc.abstractmethod
+    def save_skos(self, path):
+        """Save the contents of the subject corpus into a SKOS/Turtle
+        file with the given path name."""
+        pass  # pragma: no cover
 
 
 class SubjectIndex(metaclass=abc.ABCMeta):
