+++ tests/test_vocab_tsv.py
@@ -1,7 +1,6 @@
 """Unit tests for TSV vocabulary functionality in Annif"""
 
-from annif.corpus import SubjectFileTSV
-from annif.vocab import SubjectIndexFile
+from annif.vocab import SubjectFileTSV, SubjectIndexFile
 
 
 def test_load_tsv_uri_brackets(tmpdir):
